Move EthErc721Contract into tests dir & don't export it anymore
EthErc721Contract was only exported for use in the plasma-cash tests,
now that the tests are in this repo there's no longer any need to
export this class.
