+++ src/index.ts
@@ -36,7 +36,6 @@ export {
 } from './plasma-cash/ethereum-client'
 export { PlasmaCashTx } from './plasma-cash/plasma-cash-tx'
 export { IEthereumSigner, Web3Signer, OfflineWeb3Signer, soliditySha3 } from './solidity-helpers'
-export { EthErc721Contract } from './plasma-cash/eth-erc721-contract'
 export { Entity, IEntityParams } from './plasma-cash/entity'
 export { User as PlasmaUser } from './plasma-cash/user'
 export { SparseMerkleTree, ISparseMerkleTreeLevel } from './plasma-cash/sparse-merkle-tree'
