+++ src/tests/e2e/plasma-cash/test-hostile-operator.ts
@@ -5,7 +5,7 @@ import { runChallengeAfterDemo } from './challenge-after-demo'
 import { runChallengeBetweenDemo } from './challenge-between-demo'
 import { runChallengeBeforeDemo } from './challenge-before-demo'
 import { runRespondChallengeBeforeDemo } from './respond-challenge-before-demo'
-import { PlasmaUser } from '../../../'
+import { PlasmaUser } from '../../..'
 
 // TODO: Redeploy the Solidity contracts before each demo so the demos don't share any state.
